Adicionado botão de interface Editar na HomeConsultController + mudança de interface em tableView para editar e deletar celula

+++ CareMe/CareMe/AMVConsultCell.m
@@ -29,7 +29,7 @@
 -(void)fillWithConsult:(AMVConsult*)consult {    
     self.doctorNameLb.text = [NSString stringWithFormat:@"Dr. %@", consult.doctorName];
     self.specialityLb.text = consult.doctorSpeciality;
-    self.consultDate.text = [NSString stringWithFormat:@"%02lu/%02lu/%02lu", consult.date.day, consult.date.month, consult.date.year];
+    self.consultDate.text = [NSString stringWithFormat:@"%02lu/%02lu/%02lu", (long)consult.date.day, (long)consult.date.month, (long)consult.date.year];
     self.consultDate.textColor = [AMVCareMeUtil secondColor];
     self.doctorNameLb.textColor = [AMVCareMeUtil secondColor];
 }
